+++ CHANGELOG.md
@@ -20,6 +20,10 @@ incremented for features.
 
 * lang: Preserve all instruction data for fallback functions ([#483](https://github.com/project-serum/anchor/pull/483)).
 
+### Breaking
+
+* lang: A new `bump` keyword is added to the accounts constraints, which is used in conjunction with *both* `init` and `seeds`. When all are provided, then the `bump` is added to the seeds *and* the program executes `find_program_address` to ensure that the specified bump is the canonical bump ([#483](https://github.com/project-serum/anchor/pull/483)).
+
 ## [0.11.0] - 2021-07-03
 
 ### Features